chore: Introduce common module for test cli and server files
